Settings restart watcher improvements
Before this, the home button watcher and its receiver
were started as soon as SettingsActivity was being created,
listening for home button touches even if the user didn't change
any option that required a restart.

Now instead we start listening only if one of those options
is changed, and we stop when we are done and the reset is
triggered.

Change-Id: I9aa209083ddc56ca0679a6d807670134cd21dc69
Signed-off-by: DennySPb <dennyspb@gmail.com>
Signed-off-by: spezi77 <spezi7713@gmx.net>